Compile times.
dxcapi.h moved out of the gpu_resources header in exchange of adding it to the pch.
Shader resource moved out of the gpu_resources, it is not required in so many places.
Crash tracker needs to include dxcapi but has less depednency in the project.
Overall, no change in compile times. But I prefer lightweight headers more.