DXR first part.
Porting D3D12RaytracingSimpleLighting Microsoft sample to the project. Heap maintenance has to be rewritten so that it is compatible with my application. Resource creation, naming convention and the scene stays as in the sample for now.
This version compiles but crashes. Initialiation is not done yet, last step is acceleration structures creation in the first frame.